+++ src/widgets/index.tsx
@@ -1,6 +1,7 @@
 import Button from './Button';
 import Card from './Card';
 import CenterInfoLine from './CenterInfoLine';
+import Circles from './Circles';
 import Input from './Input';
 import Menu from './Menu';
 import MenuItem from './MenuItem';
@@ -10,5 +11,5 @@ import QuoteSlideItem from './QuoteSlideItem';
 import Title from './Title';
 
 
-export { Button, Card, CenterInfoLine, Input, Menu, MenuItem, Paragraph, PhoneNumber, QuoteSlideItem, Title };
+export { Button, Card, CenterInfoLine, Circles, Input, Menu, MenuItem, Paragraph, PhoneNumber, QuoteSlideItem, Title };